requirements.txt
timeago
python-dotenv
google-generativeai
beautifulsoup4
telethon